# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/emergency-broadcasts": {
      "get": {
        "operationId": "EmergencyBroadcastController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "meta"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmergencyBroadcastDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List emergency broadcasts (filter by status/search).",
        "tags": [
          "emergency-broadcasts"
        ]
      },
      "post": {
        "operationId": "EmergencyBroadcastController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmergencyBroadcastDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyBroadcastDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an emergency broadcast (DRAFT).",
        "tags": [
          "emergency-broadcasts"
        ]
      }
    },
    "/emergency-broadcasts/quick-text": {
      "post": {
        "operationId": "EmergencyBroadcastController_quickText",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickTextBroadcastDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyBroadcastDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create and immediately activate a text emergency broadcast.",
        "tags": [
          "emergency-broadcasts"
        ]
      }
    },
    "/emergency-broadcasts/{id}": {
      "get": {
        "operationId": "EmergencyBroadcastController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyBroadcastDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Emergency broadcast detail (with validation + affected screens).",
        "tags": [
          "emergency-broadcasts"
        ]
      },
      "patch": {
        "operationId": "EmergencyBroadcastController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmergencyBroadcastDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyBroadcastDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an emergency broadcast.",
        "tags": [
          "emergency-broadcasts"
        ]
      }
    },
    "/emergency-broadcasts/{id}/validate": {
      "get": {
        "operationId": "EmergencyBroadcastController_validate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validation report (errors, warnings, affected screen count).",
        "tags": [
          "emergency-broadcasts"
        ]
      }
    },
    "/emergency-broadcasts/{id}/activate": {
      "post": {
        "operationId": "EmergencyBroadcastController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyBroadcastDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "emergency-broadcasts"
        ]
      }
    },
    "/emergency-broadcasts/{id}/pause": {
      "post": {
        "operationId": "EmergencyBroadcastController_pause",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyBroadcastDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "emergency-broadcasts"
        ]
      }
    },
    "/emergency-broadcasts/{id}/end": {
      "post": {
        "operationId": "EmergencyBroadcastController_end",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyBroadcastDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "emergency-broadcasts"
        ]
      }
    },
    "/emergency-broadcasts/{id}/archive": {
      "post": {
        "operationId": "EmergencyBroadcastController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyBroadcastDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "emergency-broadcasts"
        ]
      }
    },
    "/emergency-broadcasts/{id}/targets": {
      "post": {
        "operationId": "EmergencyBroadcastController_addTarget",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyTargetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "emergency-broadcasts"
        ]
      }
    },
    "/emergency-broadcasts/{id}/targets/{targetId}": {
      "delete": {
        "operationId": "EmergencyBroadcastController_removeTarget",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "emergency-broadcasts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wizer Signage API",
    "description": "Wizer Signage — multi-tenant digital signage SaaS platform REST API.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 25
          },
          "total": {
            "type": "number",
            "example": 137,
            "description": "Total matching rows, ignoring pagination."
          },
          "totalPages": {
            "type": "number",
            "example": 6
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "BroadcastTargetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "targetType": {
            "type": "string",
            "enum": [
              "SCREEN",
              "SCREEN_GROUP",
              "LOCATION",
              "COMPANY"
            ]
          },
          "targetId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "targetType",
          "targetId"
        ]
      },
      "EmergencyBroadcastDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "broadcastType": {
            "type": "string",
            "enum": [
              "TEXT",
              "URL",
              "CONTENT",
              "PLAYLIST"
            ],
            "description": "Which of message/url/contentId/playlistId carries the payload."
          },
          "message": {
            "type": "object",
            "nullable": true,
            "description": "Body for a TEXT broadcast."
          },
          "url": {
            "type": "object",
            "nullable": true
          },
          "contentId": {
            "type": "object",
            "nullable": true
          },
          "playlistId": {
            "type": "object",
            "nullable": true
          },
          "priority": {
            "type": "number",
            "description": "Outranks every schedule while the broadcast is ACTIVE."
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "PAUSED",
              "ENDED",
              "ARCHIVED"
            ]
          },
          "startAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "stoppedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BroadcastTargetDto"
            }
          },
          "createdById": {
            "type": "object",
            "nullable": true
          },
          "updatedById": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "broadcastType",
          "priority",
          "status",
          "targets",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateEmergencyBroadcastDto": {
        "type": "object",
        "properties": {}
      },
      "QuickTextBroadcastDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateEmergencyBroadcastDto": {
        "type": "object",
        "properties": {}
      },
      "EmergencyTargetDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}